Moved "Utils" to utils package.
