+++ src/main/java/it/unitn/ds1/utils/WriteId.java
@@ -16,7 +16,7 @@ public class WriteId {
      * Checks if this id is precedent or equal to other
      */
     public boolean isPriorOrEqualTo(WriteId other) {
-        return this.epoch < other.epoch || this.epoch == other.epoch && this.index <= other.index;
+        return this.epoch < other.epoch || this.epoch.equals(other.epoch) && this.index <= other.index;
     }
 
     @Override
